Places hard attention modules in separate module
No other changes.
